Replace pygerrit with our own client
Instead of messing around with pygerrit internals to
make sure that client connections and streaming
threads start and stop together properly, this
change implements our own client.

This wraps the stream thread inside the gerrit client,
so that we're not trying to deal with them separately.

There are other bug fixes:

  * Keepalive connections by default
  * Client persistence when reloading configuration fixed
  * Simple JSON->object parser
  * Actions changed to refer to the properties as named
    by the JSON supplied by gerrit
  * Errors while running actions are logged and do not
    cause zoidberg to exit